Refactored BasicCalendar.tsx - separated logic to a new file - UseBasicCalendarLogic.ts, in BasicCalendar.tsx left only the view.
