+++ src/components/Accordion/Accordion.css
@@ -1,9 +1,7 @@
-/* Add this class to the parent container of the Accordion */
 .accordion-container {
     position: relative;
 }
 
-/* Styling for the Accordion */
 .accordion {
     border: 1px solid #ccc;
     margin-bottom: 3px;
