+++ src/components/Calendar/Calendar.tsx
@@ -7,6 +7,17 @@ import moment from 'moment';
 import {MyEvent} from "../BasicCalendar/BasicCalendar";
 import React from "react";
 
+moment.locale('pl', {
+    week: {
+        dow: 1,
+        doy: 1,
+    },
+});
+
+moment.updateLocale('pl', {
+    weekdaysShort: ['Niedziela', 'Poniedziałek', 'Wtorek', 'Środa', 'Czwartek', 'Piątek', 'Sobota'],
+});
+
 export interface CalendarPropsWithMyEvents extends Omit<CalendarProps<MyEvent>, 'localizer'> {}
 
 const localizer = momentLocalizer(moment)
